Favorite Music Website
After spending two days scouring the web and joining 12 different discord coding groups I haven't yet found an answer to my questions.  Do I need to link my static pages in ASP .NET CORE to my index.cshtml or is it that the order in which they are all placed in the "pipeline" effects whether or not they will load?